Fixes #662 (incl. item 3, code revision)
