+++ src/lu/fisch/structorizer/locales/de.txt
@@ -643,41 +643,6 @@ Editor.popupDisable.text=Deaktivieren/Aktivieren
 // Debug
 Editor.popupBreakpoint.text=Haltepunkt setzen/löschen
 Editor.popupBreakTrigger.text=Haltepunktauslösung festlegen...
-// Arranger index
-Editor.popupIndexGet.text=Diagram holen
-Editor.popupIndexAttributes.text=Eigenschaften anzeigen ...
-Editor.popupIndexInfo.text=Diagramm-/Gruppen-Info ...
-Editor.popupIndexSave.text=Änderungen sichern
-Editor.popupIndexRemove.text=Entfernen
-Editor.popupIndexCovered.text=Testabdeckung an/aus
-Editor.popupIndexRemoveAll.text=Alle entfernen
-Editor.popupIndexGroup.text=Gruppe erzeugen ...
-Editor.popupIndexExpandGroup.text=Gruppe erweitern ...
-Editor.popupIndexDissolve.text=Gruppe auflösen
-Editor.popupIndexDetach.text=Aus Gruppe entfernen
-Editor.popupIndexAttach.text=In Gruppe einfügen/verschieben ...
-Editor.lblSelectTargetGroup.text=Zielgruppe auswählen:
-Editor.lblGroups.text=Zugeordnete Gruppen
-Editor.lblSubroutines.text=Aufgerufene Unterprogramme
-Editor.lblIncludables.text=Einbezogene Diagramme
-Editor.lblStaleReferences.text=Unaufgelöste Diagrammverweise
-Editor.lblModifications.text=Ungesicherte Änderungen
-Editor.lblExternSubroutines.text=Verweise auf externe @oe
-Editor.lblExternIncludables.text=Verweise auf externe @ps
-Editor.btnShowGroup.tooltip=Individuelle Sichtbarkeit dieser Gruppe im Arranger
-Editor.btnGroupColors.*.tooltip=Setzt die Farbe dieser Schaltfläche als Farbe der Gruppe
-Editor.msgDefaultGroupName.text=(Auffanggruppe)
-Editor.msgGroupsAndRootsSelected.text=Sowohl Gruppen als auch Diagramme ausgewählt. Gleichzeitiges Löschen auf beiden Ebenen könnte unerwartete Ergebnisse haben.
-Editor.msgDeleteGroupMembers.text=% Gruppen werden entfernt.\n\nDarin enthaltene Diagramme bleiben erhalten, soweit sie in weieteren Gruppen vorkommen.\n\nWas soll mit Diagrammen geschehen, die nicht in weiteren Gruppen enthalten sind: Aus Arranger entfernen?\n(Anderenfalls werden sie in die Auffanggruppe übernommen.)
-Editor.msgConfirmDeleteRoots.text=Sie haben % Diagramm(e) zum Löschen ausgewählt\n\nSollen sie ggf. auch aus allen anderen Gruppen insgesamt entfernt werden?\n(Anderenfalls werden sie einfach nur aus der jeweiligen Gruppe gelöscht.)
-Editor.msgAttachOptions.0.text=Zur Gruppe hinzufügen
-Editor.msgAttachOptions.1.text=In die Gruppe verschieben
-Editor.msgAttachOptions.2.text=Abbrechen
-Editor.msgNumberOfSharedMembers.text=% Diagramme gemeinsam mit anderen Gruppen
-Editor.msgMembersIncomplete.text=Gruppe ist unvollständig: Verweise auf %1 Diagramme außerhalb der Gruppe; %2 nicht auflösbare Verweise
-Editor.msgMembersComplete.text=Gruppe ist vollständig: Keine Verweise nach außen
-Editor.msgGroupMembersChanged.text=Die Menge der enthaltenen Diagramme hat sich geändert.
-Editor.msgGroupMembersMoved.text=Die Koordinaten einiger enthaltener Diagramme wurden verändert.
 
 -----[ InputBox ]-----
 InputBox.title=Element bearbeiten
@@ -1148,6 +1113,8 @@ Arranger.btnPinDiagram.tooltip=Ausgewählte Diagramme anheften, um sie gegen Ers
 Arranger.btnSetCovered.tooltip=Unterprogrammdiagrame der Auswahl für Routinenaufrufe als testabgedeckt markieren.
 Arranger.btnRemoveDiagram.tooltip=Entfernt die ausgewählten Diagramme - Umschalttaste gedrückt halten, um alle Diagramme zu entfernen.
 Arranger.btnZoom.tooltip=Umschalttaste gedrückt halten, um heranzuzoomen.
+Arranger.chkDrawGroups.text=Zeige Gruppen
+Arranger.chkSelectGroups.text=Gruppen selektieren
 Arranger.popupHitList.text=Berührte Diagramme / Gruppen
 Arranger.popupRemove.text=Gewählte Diagramme entfernen
 Arranger.popupRemoveAll.text=Alle Diagramme entfernen
@@ -1224,10 +1191,46 @@ Surface.msgArrangementNotLoaded.text=Laden der Anordnungsdatei abgebrochen.
 Surface.msgSaveAsNewGroup.text=als neue Gruppe
 Surface.msgSelectGroup.text=Bitte festlegen, ob die Datei einer der bestehenden Gruppen aktualisiert oder eine neue Anordnung gespeichert werden soll:
 Surface.msgConfirmRemoveGroup.text=Gruppe «%» ist geleert worden. Soll sie jetzt gelöscht werden?
-Surface.msgSaveGroupChanges.text=Gruppe «%» hat ungesicherte Änderungen.\nSollen diese Änderungen gespeichert werden, bevor die Gruppe entfernt wird?
+Surface.msgSaveGroupChanges.text=Gruppe «%» hat ungesicherte Änderungen.\nSollen diese Änderungen jetzt gespeichert werden?
 Surface.msgUnsavedGroups.text=Folgende Gruppen konnten nicht gesichert werden:
 Surface.msgDiagram.text=Diagramm «%»
 Surface.msgGroup.text=Gruppe «%»
+
+----[ ArrangerIndex ]----
+ArrangerIndex.popupIndexGet.text=Diagram holen
+ArrangerIndex.popupIndexAttributes.text=Eigenschaften anzeigen ...
+ArrangerIndex.popupIndexInfo.text=Diagramm-/Gruppen-Info ...
+ArrangerIndex.popupIndexSave.text=Änderungen sichern
+ArrangerIndex.popupIndexRemove.text=Entfernen
+ArrangerIndex.popupIndexCovered.text=Testabdeckung an/aus
+ArrangerIndex.popupIndexRemoveAll.text=Alle entfernen
+ArrangerIndex.popupIndexGroup.text=Gruppe erzeugen ...
+ArrangerIndex.popupIndexExpandGroup.text=Gruppe erweitern ...
+ArrangerIndex.popupIndexDissolve.text=Gruppe auflösen
+ArrangerIndex.popupIndexDetach.text=Aus Gruppe entfernen
+ArrangerIndex.popupIndexAttach.text=In Gruppe einfügen/verschieben ...
+ArrangerIndex.lblSelectTargetGroup.text=Zielgruppe auswählen:
+ArrangerIndex.lblGroups.text=Zugeordnete Gruppen
+ArrangerIndex.lblSubroutines.text=Aufgerufene Unterprogramme
+ArrangerIndex.lblIncludables.text=Einbezogene Diagramme
+ArrangerIndex.lblStaleReferences.text=Unaufgelöste Diagrammverweise
+ArrangerIndex.lblModifications.text=Ungesicherte Änderungen
+ArrangerIndex.lblExternSubroutines.text=Verweise auf externe @oe
+ArrangerIndex.lblExternIncludables.text=Verweise auf externe @ps
+ArrangerIndex.btnShowGroup.tooltip=Individuelle Sichtbarkeit dieser Gruppe im Arranger
+ArrangerIndex.btnGroupColors.*.tooltip=Setzt die Farbe dieser Schaltfläche als Farbe der Gruppe
+ArrangerIndex.msgDefaultGroupName.text=(Auffanggruppe)
+ArrangerIndex.msgGroupsAndRootsSelected.text=Sowohl Gruppen als auch Diagramme ausgewählt. Gleichzeitiges Löschen auf beiden Ebenen könnte unerwartete Ergebnisse haben.
+ArrangerIndex.msgDeleteGroupMembers.text=% Gruppen werden entfernt.\n\nDarin enthaltene Diagramme bleiben erhalten, soweit sie in weieteren Gruppen vorkommen.\n\nWas soll mit Diagrammen geschehen, die nicht in weiteren Gruppen enthalten sind: Aus Arranger entfernen?\n(Anderenfalls werden sie in die Auffanggruppe übernommen.)
+ArrangerIndex.msgConfirmDeleteRoots.text=Sie haben % Diagramm(e) zum Löschen ausgewählt\n\nSollen sie ggf. auch aus allen anderen Gruppen insgesamt entfernt werden?\n(Anderenfalls werden sie einfach nur aus der jeweiligen Gruppe gelöscht.)
+ArrangerIndex.msgAttachOptions.0.text=Zur Gruppe hinzufügen
+ArrangerIndex.msgAttachOptions.1.text=In die Gruppe verschieben
+ArrangerIndex.msgAttachOptions.2.text=Abbrechen
+ArrangerIndex.msgNumberOfSharedMembers.text=% Diagramme gemeinsam mit anderen Gruppen
+ArrangerIndex.msgMembersIncomplete.text=Gruppe ist unvollständig: Verweise auf %1 Diagramme außerhalb der Gruppe; %2 nicht auflösbare Verweise
+ArrangerIndex.msgMembersComplete.text=Gruppe ist vollständig: Keine Verweise nach außen
+ArrangerIndex.msgGroupMembersChanged.text=Die Menge der enthaltenen Diagramme hat sich geändert.
+ArrangerIndex.msgGroupMembersMoved.text=Die Koordinaten einiger enthaltener Diagramme wurden verändert.
 
 -----> Executor
 -----[ Control ]-----
