+++ src/lu/fisch/structorizer/locales/en.txt
@@ -640,41 +640,6 @@ Editor.popupDisable.text=Disable/enable
 // Debug
 Editor.popupBreakpoint.text=Toggle breakpoint
 Editor.popupBreakTrigger.text=Specify break trigger...
-// Arranger index
-Editor.popupIndexGet.text=Get diagram
-Editor.popupIndexAttributes.text=Inspect attributes ...
-Editor.popupIndexInfo.text=Diagram/group info ...
-Editor.popupIndexSave.text=Save changes
-Editor.popupIndexRemove.text=Remove
-Editor.popupIndexCovered.text=Test-covered on/off
-Editor.popupIndexRemoveAll.text=Remove All
-Editor.popupIndexGroup.text=Create group ...
-Editor.popupIndexExpandGroup.text=Expand group ...
-Editor.popupIndexDissolve.text=Dissolve group
-Editor.popupIndexDetach.text=Detach from group
-Editor.popupIndexAttach.text=Add/move to group ...
-Editor.lblSelectTargetGroup.text=Select the target group:
-Editor.lblGroups.text=Containing groups
-Editor.lblSubroutines.text=Called @os
-Editor.lblIncludables.text=Referenced @ps
-Editor.lblStaleReferences.text=Stale diagram references
-Editor.lblModifications.text=Unsaved modifications
-Editor.lblExternSubroutines.text=Referenced external @os
-Editor.lblExternIncludables.text=Referenced external @ps
-Editor.btnShowGroup.tooltip=Individual visibility of the selected group in Arranger
-Editor.btnGroupColors.*.tooltip=Change the group color to the color on this button
-Editor.msgDefaultGroupName.text=(Default Group)
-Editor.msgGroupsAndRootsSelected.text=Both groups and diagrams selected. Removing on both levels at a time may have unexpected results.
-Editor.msgDeleteGroupMembers.text=You are going to delete % groups.\n\nThose member diagrams of them that are shared by other groups will survive.\nWhat about diagrams not shared by other groups: Remove from Arranger?\n(Otherwise they would be moved to the default group.)
-Editor.msgConfirmDeleteRoots.text=You selected % diagram(s) to be removed\n\nDo you really intend to remove it/them from all groups and Arranger?\n(Otherwise it/they would just be detached from the respective group.)
-Editor.msgAttachOptions.0.text=Add to group
-Editor.msgAttachOptions.1.text=Move to group
-Editor.msgAttachOptions.2.text=Cancel
-Editor.msgNumberOfSharedMembers.text=% members shared with other groups.
-Editor.msgMembersIncomplete.text=Group is incomplete: %1 referenced diagrams outside group, %2 stale references
-Editor.msgMembersComplete.text=Group is complete: No outward references
-Editor.msgGroupMembersChanged.text=The set of member diagrams was modified.
-Editor.msgGroupMembersMoved.text=The coordinates of some member diagrams were changed.
 
 -----[ InputBox ]-----
 InputBox.title=
@@ -1144,6 +1109,8 @@ Arranger.btnPinDiagram.tooltip=Pin the selected diagrams to make them immune aga
 Arranger.btnSetCovered.tooltip=Mark the selected routine diagrams as test-covered for subroutine calls to them.
 Arranger.btnRemoveDiagram.tooltip=Removes the selected diagrams - hold Shift key pressed to remove all diagrams.
 Arranger.btnZoom.tooltip=Hold Shift key pressed to zoom in.
+Arranger.chkDrawGroups.text=Show groups
+Arranger.chkSelectGroups.text=Select groups
 Arranger.popupHitList.text=Hit diagrams / groups
 Arranger.popupRemove.text=Remove selected diagrams
 Arranger.popupRemoveAll.text=Remove all diagrams
@@ -1220,10 +1187,46 @@ Surface.msgArrangementNotLoaded.text=Arrangement loading cancelled.
 Surface.msgSaveAsNewGroup.text=as new group
 Surface.msgSelectGroup.text=Please decide whether to update the file of an existing group or to create a new arrangement:
 Surface.msgConfirmRemoveGroup.text=Group «%» became empty. Do you want to remove it now?
-Surface.msgSaveGroupChanges.text=Group «%» has pending changes.\nDo you want to save these changes before the group is removed?
+Surface.msgSaveGroupChanges.text=Group «%» has pending changes.\nDo you want to save these changes now?
 Surface.msgUnsavedGroups.text=Couldn't save these groups (arrangements):
 Surface.msgDiagram.text=diagram «%»
 Surface.msgGroup.text=group «%»
+
+-----[ ArrangerIndex ]-----
+ArrangerIndex.popupIndexGet.text=Get diagram
+ArrangerIndex.popupIndexAttributes.text=Inspect attributes ...
+ArrangerIndex.popupIndexInfo.text=Diagram/group info ...
+ArrangerIndex.popupIndexSave.text=Save changes
+ArrangerIndex.popupIndexRemove.text=Remove
+ArrangerIndex.popupIndexCovered.text=Test-covered on/off
+ArrangerIndex.popupIndexRemoveAll.text=Remove All
+ArrangerIndex.popupIndexGroup.text=Create group ...
+ArrangerIndex.popupIndexExpandGroup.text=Expand group ...
+ArrangerIndex.popupIndexDissolve.text=Dissolve group
+ArrangerIndex.popupIndexDetach.text=Detach from group
+ArrangerIndex.popupIndexAttach.text=Add/move to group ...
+ArrangerIndex.lblSelectTargetGroup.text=Select the target group:
+ArrangerIndex.lblGroups.text=Containing groups
+ArrangerIndex.lblSubroutines.text=Called @os
+ArrangerIndex.lblIncludables.text=Referenced @ps
+ArrangerIndex.lblStaleReferences.text=Stale diagram references
+ArrangerIndex.lblModifications.text=Unsaved modifications
+ArrangerIndex.lblExternSubroutines.text=Referenced external @os
+ArrangerIndex.lblExternIncludables.text=Referenced external @ps
+ArrangerIndex.btnShowGroup.tooltip=Individual visibility of the selected group in Arranger
+ArrangerIndex.btnGroupColors.*.tooltip=Change the group color to the color on this button
+ArrangerIndex.msgDefaultGroupName.text=(Default Group)
+ArrangerIndex.msgGroupsAndRootsSelected.text=Both groups and diagrams selected. Removing on both levels at a time may have unexpected results.
+ArrangerIndex.msgDeleteGroupMembers.text=You are going to delete % groups.\n\nThose member diagrams of them that are shared by other groups will survive.\nWhat about diagrams not shared by other groups: Remove from Arranger?\n(Otherwise they would be moved to the default group.)
+ArrangerIndex.msgConfirmDeleteRoots.text=You selected % diagram(s) to be removed\n\nDo you really intend to remove it/them from all groups and Arranger?\n(Otherwise it/they would just be detached from the respective group.)
+ArrangerIndex.msgAttachOptions.0.text=Add to group
+ArrangerIndex.msgAttachOptions.1.text=Move to group
+ArrangerIndex.msgAttachOptions.2.text=Cancel
+ArrangerIndex.msgNumberOfSharedMembers.text=% members shared with other groups.
+ArrangerIndex.msgMembersIncomplete.text=Group is incomplete: %1 referenced diagrams outside group, %2 stale references
+ArrangerIndex.msgMembersComplete.text=Group is complete: No outward references
+ArrangerIndex.msgGroupMembersChanged.text=The set of member diagrams was modified.
+ArrangerIndex.msgGroupMembersMoved.text=The coordinates of some member diagrams were changed.
 
 -----> Executor
 -----[ Control ]-----
