+++ src/lu/fisch/structorizer/locales/es.txt
@@ -644,41 +644,6 @@ Editor.popupDisable.text=Desactivar/Reactivar
 // Breakpoint
 Editor.popupBreakpoint.text=Poner/eliminar breakpoint
 Editor.popupBreakTrigger.text=Fija la condición de parar ...
-// Arranger index
-Editor.popupIndexGet.text=Coger diagrama
-Editor.popupIndexAttributes.text=Inspiciar propiedades ...
-Editor.popupIndexInfo.text=Informaciones ...
-Editor.popupIndexSave.text=Guardar cambios
-Editor.popupIndexRemove.text=Quitar
-Editor.popupIndexCovered.text=Cambiar cobertura de test
-Editor.popupIndexRemoveAll.text=Quitar todos
-Editor.popupIndexGroup.text=Crear grupo ...
-Editor.popupIndexExpandGroup.text=Expandir grupo ...
-Editor.popupIndexDissolve.text=Descomponer grupo
-Editor.popupIndexDetach.text=Retirar del grupo
-Editor.popupIndexAttach.text=Añadir/mover a un grupo ...
-Editor.lblSelectTargetGroup.text=Selecciona el grupo destinatorio:
-Editor.lblGroups.text=Grupos asociados
-Editor.lblSubroutines.text=Subrutinas calladas
-Editor.lblIncludables.text=Incluíbles requeridos
-Editor.lblStaleReferences.text=Referencias no satisfechas
-Editor.lblModifications.text=Modificaciones no guardados
-Editor.lblExternSubroutines.text=Requeridas @os externas
-Editor.lblExternIncludables.text=Requeridos @ps externos 
-Editor.btnShowGroup.tooltip=Visibilidad individual del grupo en Arranger
-Editor.btnGroupColors.*.tooltip=Change the group color to the color on this button
-Editor.msgDefaultGroupName.text=(Grupo de incumplimiento)
-Editor.msgGroupsAndRootsSelected.text=Tanto grupos como diagramas seleccionados. Remover algo de ambos niveles a la vez puede resultar en efectos inesperados.
-Editor.msgDeleteGroupMembers.text=Se van a eliminar % grupos.\n\nAquellos diagramas miembros que pertenecen también a otros grupos se quedarán.\n¿Que hacer con diagramas no compartidos con demás grupos - quitar del Arranger?\n(De otro modo serán movidos al grupo de incumplimiento.)
-Editor.msgConfirmDeleteRoots.text=You selected % diagram(s) to be removed\n\nDo you really intend to remove it/them from all groups and Arranger?\n(Otherwise it/they would just be detached from the respective group.)
-Editor.msgAttachOptions.0.text=Añadir al grupo
-Editor.msgAttachOptions.1.text=Mover al grupo
-Editor.msgAttachOptions.2.text=Cancelar
-Editor.msgNumberOfSharedMembers.text=% diagramas miembros compartidos con otros grupos.
-Editor.msgMembersIncomplete.text=Grupo incompleto: %1 requeridos diagramas están fuera del grupo; %2 requeridos diagramas faltan
-Editor.msgMembersComplete.text=Grupo completo: No requeridos diagramas fuera del grupo
-Editor.msgGroupMembersChanged.text=El conjunto de miembros se ha cambiado.
-Editor.msgGroupMembersMoved.text=Coordenadas de algunos miembros fueron variados.
 
 -----[ InputBox ]-----
 InputBox.title=
@@ -1161,6 +1126,8 @@ Arranger.popupHelp.tooltip=Abres la página "Arranger" del online manual en el b
 Arranger.popupKeyBindings.text=Atajos de teclado ...
 Arranger.popupKeyBindings.tooltip=Muestras la lista de atajos de teclado en el browser
 Arranger.popupRearrange.text=Recolocar por grupos
+Arranger.chkDrawGroups.text=Dibujar grupos
+Arranger.chkSelectGroups.text=Selecc. grupos
 Arranger.msgConfirmRemoveAll.text=¿Quieres reálmente quitar todos los diagramas de Arranger?
 Arranger.msgTitleWarning.text=Advertencia
 Arranger.msgTitleIllegal.text=¡Operación ilícita!
@@ -1223,10 +1190,46 @@ Surface.msgArrangementNotLoaded.text=Cancelado cargar colocación.
 Surface.msgSaveAsNewGroup.text=como grupo nuevo
 Surface.msgSelectGroup.text=Decide po favor entre sobrescribir el fichero asociado a uno de los grupos existentes y crear nueva colocación:
 Surface.msgConfirmRemoveGroup.text=Grupo «%» fue vaciado. ¿Eliminarlo?
-Surface.msgSaveGroupChanges.text=Grupo «%» tiene modificaciones no grabados.\nGrabar las modificaciones antes de eliminar el grupo?
+Surface.msgSaveGroupChanges.text=Grupo «%» tiene modificaciones no grabados.\n¿Grabar las modificaciones ahora?
 Surface.msgUnsavedGroups.text=No fue posible guardar estos diagramas:
 Surface.msgDiagram.text=diagrama «%»
 Surface.msgGroup.text=grupo «%»
+
+// Arranger index
+ArrangerIndex.popupIndexGet.text=Coger diagrama
+ArrangerIndex.popupIndexAttributes.text=Inspiciar propiedades ...
+ArrangerIndex.popupIndexInfo.text=Informaciones ...
+ArrangerIndex.popupIndexSave.text=Guardar cambios
+ArrangerIndex.popupIndexRemove.text=Quitar
+ArrangerIndex.popupIndexCovered.text=Cambiar cobertura de test
+ArrangerIndex.popupIndexRemoveAll.text=Quitar todos
+ArrangerIndex.popupIndexGroup.text=Crear grupo ...
+ArrangerIndex.popupIndexExpandGroup.text=Expandir grupo ...
+ArrangerIndex.popupIndexDissolve.text=Descomponer grupo
+ArrangerIndex.popupIndexDetach.text=Retirar del grupo
+ArrangerIndex.popupIndexAttach.text=Añadir/mover a un grupo ...
+ArrangerIndex.lblSelectTargetGroup.text=Selecciona el grupo destinatorio:
+ArrangerIndex.lblGroups.text=Grupos asociados
+ArrangerIndex.lblSubroutines.text=Subrutinas calladas
+ArrangerIndex.lblIncludables.text=Incluíbles requeridos
+ArrangerIndex.lblStaleReferences.text=Referencias no satisfechas
+ArrangerIndex.lblModifications.text=Modificaciones no guardados
+ArrangerIndex.lblExternSubroutines.text=Requeridas @os externas
+ArrangerIndex.lblExternIncludables.text=Requeridos @ps externos 
+ArrangerIndex.btnShowGroup.tooltip=Visibilidad individual del grupo en Arranger
+ArrangerIndex.btnGroupColors.*.tooltip=Change the group color to the color on this button
+ArrangerIndex.msgDefaultGroupName.text=(Grupo de incumplimiento)
+ArrangerIndex.msgGroupsAndRootsSelected.text=Tanto grupos como diagramas seleccionados. Remover algo de ambos niveles a la vez puede resultar en efectos inesperados.
+ArrangerIndex.msgDeleteGroupMembers.text=Se van a eliminar % grupos.\n\nAquellos diagramas miembros que pertenecen también a otros grupos se quedarán.\n¿Que hacer con diagramas no compartidos con demás grupos - quitar del Arranger?\n(De otro modo serán movidos al grupo de incumplimiento.)
+ArrangerIndex.msgConfirmDeleteRoots.text=You selected % diagram(s) to be removed\n\nDo you really intend to remove it/them from all groups and Arranger?\n(Otherwise it/they would just be detached from the respective group.)
+ArrangerIndex.msgAttachOptions.0.text=Añadir al grupo
+ArrangerIndex.msgAttachOptions.1.text=Mover al grupo
+ArrangerIndex.msgAttachOptions.2.text=Cancelar
+ArrangerIndex.msgNumberOfSharedMembers.text=% diagramas miembros compartidos con otros grupos.
+ArrangerIndex.msgMembersIncomplete.text=Grupo incompleto: %1 requeridos diagramas están fuera del grupo; %2 requeridos diagramas faltan
+ArrangerIndex.msgMembersComplete.text=Grupo completo: No requeridos diagramas fuera del grupo
+ArrangerIndex.msgGroupMembersChanged.text=El conjunto de miembros se ha cambiado.
+ArrangerIndex.msgGroupMembersMoved.text=Coordenadas de algunos miembros fueron variados.
 
 -----> Executor
 -----[ Control ]-----
